Make convert output private. Rename json_converter -> writers.
