+++ src/python/deepgnn/graph_engine/snark/BUILD
@@ -16,7 +16,8 @@ py_library(
         "convert.py",
         "meta_merger.py",
         "converter/__init__.py",
-        "converter/json_converter.py",
+        "converter/process.py",
+        "converter/writers.py",
         "converter/options.py",
         "decoders.py",
         "dispatcher.py",
